Promote Aurora design system to shared cross-platform code
